0.3/soundshine_bot/requirements.txt
discord.py
aiohttp
python-dotenv
requests
```